Add CRUD endpoints

+++ java-api/src/main/java/com/db/grad/javaapi/repository/SecurityRepository.java
@@ -1,10 +1,9 @@
 package com.db.grad.javaapi.repository;
 
-import com.db.grad.javaapi.model.Books;
 import com.db.grad.javaapi.model.Security;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SecurityRepository extends JpaRepository<Security, Long> {
+public interface SecurityRepository extends JpaRepository<Security, Integer> {
 }
